backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
gunicorn>=21.2
python-dotenv>=1.0
dj-database-url>=2.2